requirements.txt
# MRP系统依赖包
# Python >= 3.8

# 数据处理
pandas>=1.5.0
numpy>=1.21.0
openpyxl>=3.0.0  # Excel文件读写

# 科学计算
scipy>=1.9.0

# 可视化
matplotlib>=3.5.0
seaborn>=0.11.0

# 配置文件
pyyaml>=6.0

# 日志
loguru>=0.6.0

# 开发工具
pytest>=7.0.0
black>=22.0.0
isort>=5.10.0
